refactor: Remove provider
