+++ lib/app.dart
@@ -1,9 +1,9 @@
 import 'package:admin/common/constants.dart';
-import 'package:admin/controllers/menu_app_controller.dart';
+import 'package:admin/home/cubit/menu_app_cubit.dart';
 import 'package:admin/home/home_screen.dart';
 import 'package:flutter/material.dart';
+import 'package:flutter_bloc/flutter_bloc.dart';
 import 'package:google_fonts/google_fonts.dart';
-import 'package:provider/provider.dart';
 
 class App extends StatelessWidget {
   @override
@@ -18,12 +18,16 @@ class App extends StatelessWidget {
         ).apply(bodyColor: Colors.white),
         canvasColor: secondaryColor,
       ),
-      home: MultiProvider(
-        providers: [
-          ChangeNotifierProvider(create: (context) => MenuAppController()),
-        ],
+      home: BlocProvider<MenuAppCubit>(
+        create: (context) => MenuAppCubit(),
         child: HomeScreen(),
       ),
+      // home: MultiProvider(
+      //   providers: [
+      //     ChangeNotifierProvider(create: (context) => MenuAppController()),
+      //   ],
+      //   child: HomeScreen(),
+      // ),
     );
   }
 }
