--- lib/controllers/menu_app_controller.dart
@@ -1,13 +1,0 @@
-import 'package:flutter/material.dart';
-
-class MenuAppController extends ChangeNotifier {
-  final GlobalKey<ScaffoldState> _scaffoldKey = GlobalKey<ScaffoldState>();
-
-  GlobalKey<ScaffoldState> get scaffoldKey => _scaffoldKey;
-
-  void controlMenu() {
-    if (!_scaffoldKey.currentState!.isDrawerOpen) {
-      _scaffoldKey.currentState!.openDrawer();
-    }
-  }
-}
