+++ lib/home/cubit/menu_appcubit_state.dart
@@ -1,0 +1,6 @@
+part of 'menu_app_cubit.dart';
+
+@immutable
+sealed class MenuAppState {}
+
+final class MenuAppInitial extends MenuAppState {}
