+++ pubspec.yaml
@@ -20,7 +20,7 @@ dependencies:
   google_fonts: ^6.2.1
   flutter_svg: ^2.0.10
   fl_chart: ^0.68.0
-  provider: ^6.1.2
+  flutter_bloc: ^9.1.1
 
 dev_dependencies:
   flutter_test:
